--- a/src/function/scalar/string_functions.cpp
+++ b/src/function/scalar/string_functions.cpp
@@ -11,6 +11,7 @@ void BuiltinFunctions::RegisterStringFunctions() {
 	Register<ConcatFun>();
 	Register<LengthFun>();
 	Register<LikeFun>();
+	Register<PrintfFun>();
 	Register<RegexpFun>();
 	Register<SubstringFun>();
 }