--- a/src/include/duckdb/function/scalar/string_functions.hpp
+++ b/src/include/duckdb/function/scalar/string_functions.hpp
@@ -53,6 +53,10 @@ struct SubstringFun {
 	static void RegisterFunction(BuiltinFunctions &set);
 };
 
+struct PrintfFun {
+	static void RegisterFunction(BuiltinFunctions &set);
+};
+
 struct RegexpMatchesBindData : public FunctionData {
 	RegexpMatchesBindData(std::unique_ptr<re2::RE2> constant_pattern, string range_min, string range_max,
 			      bool range_success);