… CMakeLists.txt
… benchmark/micro/cast.cpp
… src/CMakeLists.txt
… src/function/scalar/string/CMakeLists.txt
… src/include/duckdb/common/operator/cast_operators.hpp
… test/common/test_cast.cpp
… test/common/test_date.cpp
… test/sql/cast/CMakeLists.txt
… test/sql/cast/test_string_cast.cpp
… test/sql/date/CMakeLists.txt
… test/sql/date/test_date.cpp
… test/sql/date/test_time.cpp
… test/sql/date/test_timestamp.cpp
… test/sql/simple/CMakeLists.txt
… test/sql/simple/test_printf.cpp
… third_party/CMakeLists.txt
… third_party/fmt/CMakeLists.txt
… third_party/fmt/LICENSE.rst
… third_party/fmt/format.cc
… third_party/fmt/include/fmt/core.h
… third_party/fmt/include/fmt/format-inl.h
… third_party/fmt/include/fmt/format.h
… third_party/fmt/include/fmt/ostream.h
… third_party/fmt/include/fmt/printf.h