This PR adds optimized casting from integers, dates, times and timestamps to strings. Previously, this would simply use the C++ standard library to_string methods, which would create a std::string object that then needed to be copied. Now we first figure out the length of the converted string, allocate space for it, and then directly perform the conversion within the target buffer. This makes the conversions much faster:

INT -> STRING Benchmark:
0.65s -> 0.25s

DATE -> STRING Benchmark:
1.8s ->  0.37s

TIME -> STRING Benchmark:
1.65s -> 0.24s

TIMESTAMP -> STRING Benchmark
4.2s -> 0.7s

This in turn also speeds up things like CSV export where to string conversions are used.